improve output (less bugs), abstract the Run function to not require arg parsing
