+++ pkg/consolewriter/consolewriter.go
@@ -102,6 +102,15 @@ func (c *ConsoleWriter) Printf(format string, v ...interface{}) {
 	}
 }
 
+// Print calls l.Output to print to the logger.
+// Arguments are handled in the manner of fmt.Print.
+func (c *ConsoleWriter) Print(v ...interface{}) {
+	err := c.Output(2, fmt.Sprint(v...))
+	if err != nil {
+		fmt.Println(err)
+	}
+}
+
 //Fprintf is a helper function, will write the format (time etc) using the the supplied writer
 func (c *ConsoleWriter) Fprintf(w io.Writer, format string, a ...interface{}) (n int, err error) {
 	now := time.Now() // get this early.
